# sql: ON CONFLICT DO UPDATE form of upsert