--- a/sql/select_qvalue.go
+++ b/sql/select_qvalue.go
@@ -44,8 +44,8 @@ func (cr columnRef) get() ResultColumn {
 }
 
 type qvalResolver struct {
-	table *tableInfo
-	qvals qvalMap
+	tables []*tableInfo
+	qvals  qvalMap
 }
 
 // findColumn looks up the column described by a QualifiedName. The qname will be normalized.
@@ -63,22 +63,25 @@ func (qt qvalResolver) findColumn(qname *parser.QualifiedName) (columnRef, error
 		return ref, err
 	}
 
-	// TODO(radu): when we support multiple FROMs, we will find the node with the correct alias; if
-	// no alias is given, we will search for the column in all FROMs and make sure there is only
-	// one.  For now we just check that the name matches (if given).
-	if qname.Base == "" || sqlbase.EqualName(qt.table.alias, string(qname.Base)) {
-		colName := sqlbase.NormalizeName(qname.Column())
-		for idx, col := range qt.table.columns {
-			if sqlbase.NormalizeName(col.Name) == colName {
-				ref.table = qt.table
-				ref.colIdx = idx
-				return ref, nil
+	colName := sqlbase.NormalizeName(qname.Column())
+	for _, table := range qt.tables {
+		if qname.Base == "" || sqlbase.EqualName(table.alias, string(qname.Base)) {
+			for idx, col := range table.columns {
+				if sqlbase.NormalizeName(col.Name) == colName {
+					if ref.colIdx != invalidColIdx {
+						return ref, fmt.Errorf("column reference \"%s\" is ambiguous", qname)
+					}
+					ref.table = table
+					ref.colIdx = idx
+				}
 			}
 		}
 	}
 
-	err := fmt.Errorf("qualified name \"%s\" not found", qname)
-	return ref, err
+	if ref.colIdx == invalidColIdx {
+		return ref, fmt.Errorf("qualified name \"%s\" not found", qname)
+	}
+	return ref, nil
 }
 
 // qvalue implements the parser.VariableExpr interface and is used as a
@@ -224,13 +227,15 @@ func (s *selectNode) resolveQNames(expr parser.Expr) (parser.Expr, error) {
 	if s.planner != nil {
 		v = &s.planner.qnameVisitor
 	}
-	return resolveQNames(expr, &s.table, s.qvals, v)
+	return resolveQNames(expr, []*tableInfo{&s.table}, s.qvals, v)
 }
 
 // resolveQNames walks the provided expression and resolves all qualified
 // names using the tableInfo and qvalMap. The function takes an optional
 // qnameVisitor to provide the caller the option of avoiding an allocation.
-func resolveQNames(expr parser.Expr, table *tableInfo, qvals qvalMap, v *qnameVisitor) (parser.Expr, error) {
+func resolveQNames(
+	expr parser.Expr, tables []*tableInfo, qvals qvalMap, v *qnameVisitor,
+) (parser.Expr, error) {
 	if expr == nil {
 		return expr, nil
 	}
@@ -239,20 +244,22 @@ func resolveQNames(expr parser.Expr, table *tableInfo, qvals qvalMap, v *qnameVi
 	}
 	*v = qnameVisitor{
 		qt: qvalResolver{
-			table: table,
-			qvals: qvals,
+			tables: tables,
+			qvals:  qvals,
 		},
 	}
 	expr, _ = parser.WalkExpr(v, expr)
 	return expr, v.err
 }
 
-// Populates the datum fields in the qval map given a row of values.
-func (q qvalMap) populateQVals(row parser.DTuple) {
+// Populates one table's datum fields in the qval map given a row of values.
+func (q qvalMap) populateQVals(table *tableInfo, row parser.DTuple) {
 	for ref, qval := range q {
-		qval.datum = row[ref.colIdx]
-		if qval.datum == nil {
-			panic(fmt.Sprintf("Unpopulated value for column %d", ref.colIdx))
+		if ref.table == table {
+			qval.datum = row[ref.colIdx]
+			if qval.datum == nil {
+				panic(fmt.Sprintf("Unpopulated value for column %d", ref.colIdx))
+			}
 		}
 	}
 }