… sql/parser/parse_test.go
… sql/parser/sql.go
… sql/testdata/upsert
… sql/update.go
… sql/upsert.go