@RaduBerinde 2016-05-10
Great stuff Dan!  Some initial comments

<details><summary><i>Previously, paperstreet (Dan Harrison) wrote…</i></summary>
> cc @knz @nvanbenschoten @RaduBerinde </details>


---

Review status: 0 of 13 files reviewed at latest revision, 8 unresolved discussions, all commit checks successful.

---

*[sql/select_qvalue.go, line 77 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQADCO2sK8KpEcOmNK:-KHQADCO2sK8KpEcOmNL:1937386386) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/select_qvalue.go#L77)):*
> ```Go
> 		}
> 	} else {
> 		for _, table := range qt.aliasedTables {
> ```

It seems like `aliasedTables` acts as a "fallback". If the actual table name is `excluded` what do we want to  happen? (please add a test for that)

The semantics of `aliasedTables` do not match how we want this code to behave when we have multiple tables in a select. What we would want is described in the TODO you removed. Until we actually implement that, please leave the TODO somewhere.

We could implement this to satisfy both usecases:
 - don't add `aliasedTables`; change `table` to a slice
 - add a `defaultTableIdx` field

If `defaultTableIdx` is set, the code prefers the default table if `Base` is not set or if it matches that table. Otherwise it looks at the other tables.

If `defaultTableIdx` is not set, the code finds the table with the correct name. If Base is not given, it finds a table where the column name makes sense AND verifies that it is the ONLY table where that column name makes sense. This behavior can be left as a TODO.

---

*[sql/select_qvalue.go, line 267 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQ90IntL3TbrCNrQjO:-KHQ90IntL3TbrCNrQjP:534435818) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/select_qvalue.go#L267)):*
> ```Go
> func (q qvalMap) populateQVals(row parser.DTuple) {
> 	for ref, qval := range q {
> 		qval.datum = row[ref.colIdx+ref.table.columnOffset]
> ```

Why not just call `populateQvals(row[offset:])` instead of adding the `columnOffset`?

---

*[sql/tablewriter.go, line 293 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQCw4ui5Y6o8YNHAtL:-KHQCw4ui5Y6o8YNHAtM:1046139874) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/tablewriter.go#L293)):*
> ```Go
> 		for i, result := range b.Results {
> 			if len(result.Rows) == 0 {
> 				// No conflict for this row, so leave upsertRowPKs[i] as nil.
> ```

Can remove the first `if` and just leave the comment. Also, maybe panic if we get more than 1 row?

---

*[sql/tablewriter.go, line 326 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQDCEAqUeS3CkxDz91:-KHQDCEAqUeS3CkxDz92:-1836978434) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/tablewriter.go#L326)):*
> ```Go
> 			pkSpans = append(pkSpans, sqlbase.Span{Start: primaryKey, End: primaryKey.PrefixEnd()})
> 			if _, ok := rowIdxForPrimaryKey[string(primaryKey)]; ok {
> 				return nil, fmt.Errorf("UPSERT/ON CONFLICT DO UPDATE command cannot affect row a second time")
> ```

this error message is very unclear.. maybe see what error `pg` would return in this case and match that?
Also, is this error even possible with the `UPSERT` syntax?

---

*[sql/update.go, line 346 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQ9SqSi3FqRHJZalWj:-KHQ9SqSi3FqRHJZalWk:1040615641) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/update.go#L346)):*
> ```Go
> }
> 
> func (p *planner) namesForExprs(exprs parser.UpdateExprs) (parser.QualifiedNames, error) {
> ```

function needs comment

---

*[sql/upsert.go, line 103 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQ7xB7Irc_-54U1uzz:-KHQ7xB7Irc_-54U1v--:128717271) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/upsert.go#L103)):*
> ```Go
> }
> 
> func upsertExprsAndIndex(
> ```

Function needs a description

---

*[sql/upsert.go, line 163 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQ87CsGRXcBOjRwGQO:-KHQ87CsGRXcBOjRwGQP:-919091961) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/upsert.go#L163)):*
> ```Go
> 	}
> 	if conflictIndex == nil {
> 		return nil, nil, util.Errorf("there is no unique or exclusion constraint matching the ON CONFLICT specification")
> ```

Isn't this a user-facing error? (if yes, use `fmt.Errorf`)

---

*[sql/testdata/upsert, line 68 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQDtX1iR7G_54Ow0bD:-KHQDtX1iR7G_54Ow0bE:-1109991100) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/testdata/upsert#L68)):*
> ```
> 
> statement ok
> INSERT INTO abc VALUES (1, 2, 3) ON CONFLICT (c) DO UPDATE SET a = 4
> ```

are we no longer testing the `UPSERT` form?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-:-KHQDyLwhDHFjmJPCbow:-820105656)*
<!-- Sent from Reviewable.io -->