@nvanbenschoten 2016-05-10
Nice! Reviewed `sql` subpackages so far.

<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> Great stuff Dan!  Some initial comments</details>


---

Reviewed 7 of 13 files at r1.
Review status: 7 of 13 files reviewed at latest revision, 10 unresolved discussions, some commit checks failed.

---

*[sql/parser/insert.go, line 76 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQfSyjPWGskEGeVu6L:-KHQfSyjPWGskEGeVu6M:102871454) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/parser/insert.go#L76)):*
> ```Go
> //
> // The zero value for OnConflict is used to signal the UPSERT short form, which
> // uses the primary key for Index and the values being inserted for Exprs.
> ```

Does this comment need to be updated? It looks like `Index` is gone.

Also maybe add *exprs* and *wheres* to the `ON CONFLICT index DO UPDATE SET` part of the comment to show how they fit in.

---

*[sql/parser/insert.go, line 85 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQfoW1wMUREszbaVEw:-KHQfoW1wMUREszbaVEx:162958664) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/parser/insert.go#L85)):*
> ```Go
> // IsUpsertAlias returns true if the UPSERT syntactic sugar was used.
> func (oc *OnConflict) IsUpsertAlias() bool {
> 	return oc.Columns == nil && oc.Exprs == nil && oc.Where == nil
> ```

Consider just comparing agains the zero value. `return oc != nil && *oc == OnConflict{}`

That would also pull in the nil check, which I think cleans up the semantics a bit and allows you to remove it at least for the first `if node.OnConflict != nil && node.OnConflict.IsUpsertAlias() {`

---

*[sql/testdata/upsert, line 68 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-KHQDtX1iR7G_54Ow0bD:-KHQgRSNJm-gJxY5P-HM:15559160) ([raw file](https://github.com/cockroachdb/cockroach/blob/5bfb714375fba0b6e3a82ffa4e2c8d209bd3b87d/sql/testdata/upsert#L68)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> are we no longer testing the `UPSERT` form?</details>
I think @paperstreet pulled them up top. Still, it would be nice to double check that we're not losing any logical test coverage here.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6591#-:-KHQhaJpfJVMK1UVHP5C:-359669390)*
<!-- Sent from Reviewable.io -->